+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `Grid` and `GridObject` for the scene objects.
+* Added `TagObject`, `LineObject` and `PointObject` for the scene objects.
 * Added test files for configurations.
 * Added doc build, and many doc-building syntax fixes.
 * Added `scene.json` for storing the scene configurations. Mostly store the default appearance features of the objects.